Fixed permissions system, adding and assigning users to collections

+++ backend/api/collection.py
@@ -5,9 +5,10 @@ from fastapi.security import OAuth2PasswordBearer
 from pydantic import BaseModel
 from sqlalchemy.orm.exc import NoResultFound
 
-from api.token import decode_access_token
+from api.token import decode_access_token, check_permissions
 from repository.collection import Collection, CollectionRepository
 from repository.item import Item, ItemRepository
+from repository.permission import PermissionRepository
 
 router = APIRouter()
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
@@ -18,12 +19,18 @@ class CreateCollectionRequest(BaseModel):
     description: Optional[str]
 
 
+class AssignRequest(BaseModel):
+    user_id: int
+    permission: str
+
+
 @router.get(
     "/collections/{collection_id}",
     tags=['collections'],
     responses={404: {'detail': 'Collection not found'}}
 )
 async def get_collection(collection_id: int, token: str = Depends(oauth2_scheme)) -> Collection:
+    check_permissions(token=token, required_level='User', collection_id=collection_id)
     try:
         return CollectionRepository.get_collection_by_id(collection_id=collection_id)
     except NoResultFound:
@@ -48,6 +55,7 @@ async def get_collections_for_user(user_id: int, token: str = Depends(oauth2_sch
     responses={404: {'detail': 'Collection not found'}}
 )
 async def get_collection_items(collection_id: int, token: str = Depends(oauth2_scheme)) -> list[Item]:
+    check_permissions(token=token, required_level='User', collection_id=collection_id)
     try:
         return ItemRepository.get_items_for_collection_id(collection_id=collection_id)
     except NoResultFound:
@@ -72,9 +80,9 @@ async def list_collections(token: str = Depends(oauth2_scheme)) -> list[Collecti
     tags=['collections'],
     responses={400: {'detail': 'Invalid request payload'}}
 )
-async def create_collection(collection: CreateCollectionRequest, token: str = Depends(oauth2_scheme)) -> None:
+async def create_collection(collection: CreateCollectionRequest, token: str = Depends(oauth2_scheme)) -> int:
     user_id = int(decode_access_token(token).sub)
-    CollectionRepository.create_collection(collection.dict(), user_id)
+    return CollectionRepository.create_collection(collection.dict(), user_id)
 
 
 @router.delete(
@@ -83,7 +91,19 @@ async def create_collection(collection: CreateCollectionRequest, token: str = De
     responses={404: {'detail': 'Collection not found'}}
 )
 async def delete_collection(collection_id: int, token: str = Depends(oauth2_scheme)) -> None:
+    check_permissions(token=token, required_level='Collection Administrator', collection_id=collection_id)
     try:
         return CollectionRepository.delete_collection(collection_id)
     except NoResultFound:
         raise HTTPException(status_code=404, detail='Collection not found') from None
+
+
+@router.post(
+    '/collections/{collection_id}/users',
+    tags=['collections', 'users'],
+    responses={404: {'detail': 'User or collection not found'}}
+)
+async def assign_user(collection_id: int, assign_request: AssignRequest, token: str = Depends(oauth2_scheme)) -> None:
+    check_permissions(token=token, required_level='Collection Administrator', collection_id=collection_id)
+    permission_repo = PermissionRepository()
+    permission_repo.assign_user(assign_request.user_id, collection_id, assign_request.permission)
